Backend/user_service/requirements.txt
fastapi==0.124.0
sqlalchemy==2.0.43
PyJWT==2.10.1
psycopg2-binary==2.9.9
pika==1.3.2
pydantic==2.11.10
pydantic-settings==2.0.3
redis==6.4.0
alembic==1.17.2
granian[reload]==2.6.0
python-dotenv==1.1.1
pytest==8.3.4
httpx==0.27.2